Issue #77 * Renamed "Crossed" to "InsertionIterator". * Renamed "Conjecturer" to "InsertionStrategy".
