+++ tests/test_algorithms/test_heuristics/test_insertion/iterators/test_best_stateless.py
@@ -8,23 +8,28 @@ from tests.utils import (
 )
 
 
-class TestStatelessCrosser(unittest.TestCase):
+class TestStatelessInsertionIterator(unittest.TestCase):
 
     def test_creation(self):
         job = jit.Job(generate_trips(10), objective_cls=jit.DialARideObjective)
         fleet = jit.Fleet(generate_vehicles(10))
-        dispatcher = jit.StatelessCrosser(
+        randomized_size = 2
+        dispatcher = jit.BestStatelessInsertionIterator(
+            randomized_size=randomized_size,
             job=job,
             fleet=fleet,
         )
         self.assertEqual(job, dispatcher.job)
         self.assertEqual(fleet, dispatcher.fleet)
         self.assertEqual(len(job.trips), len(list(dispatcher.iterator)))
+        self.assertEqual(dispatcher.randomized_size, randomized_size)
 
     def test_flush(self):
         job = jit.Job(generate_trips(10), objective_cls=jit.DialARideObjective)
         fleet = jit.Fleet(generate_vehicles(10))
-        dispatcher = jit.StatelessCrosser(
+        randomized_size = 2
+        dispatcher = jit.BestStatelessInsertionIterator(
+            randomized_size=randomized_size,
             job=job,
             fleet=fleet,
         )
