+++ tests/test_algorithms/test_heuristics/test_insertion/test_strategies/test_sampling.py
@@ -8,14 +8,14 @@ from tests.utils import (
 )
 
 
-class TestConjecturer(unittest.TestCase):
+class TestSamplingStrategy(unittest.TestCase):
 
     def test_compute(self):
-        conjecturer = jit.IntensiveConjecturer()
+        strategy = jit.SamplingInsertionStrategy()
         route = generate_one_route(planned_trips_min=10)
         trips = generate_trips(5)
 
-        conjectured_routes = conjecturer.compute(route, trips)  # noqa
+        conjectured_routes = strategy.compute(route, trips)  # noqa
         self.assertTrue(True)
         # TODO: Improve assertions
